Port InputPluginCommon to use the original IniFile code, get rid of Billiard's.
a few minor warning/compile fixes.

git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@5604 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DiscIO/Src/FileMonitor.cpp
@@ -95,7 +95,7 @@ void ReadGC(std::string FileName)
 }
 
 // Check if we should play this file
-void CheckFile(std::string File, int Size)
+void CheckFile(std::string File, u64 Size)
 {
 	// Don't do anything if the log is unselected
 	if (!LogManager::GetInstance()->isEnable(LogTypes::FILEMON)) return;
